# server/docs/openapi.yaml
openapi: 3.0.4
info:
  title: CSK Choir Hub
  description: This is the API declaration for CSK Choir Hub.
  contact:
    email: webmasters@choir.chs.chalmers.se
  version: 1.0.0
...
servers:
  - url: https://portal.chalemrssangkor.se/api
...
paths:
...
  /roles:
    get:
      tags:
        - roles
      summary: Get all roles.
      description: Gets the full list of all roles.
      operationId: getRoles
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
            application/xml:
              schema:
                type: array
                items:
                  type: object
        '400':
          description: Token missing or access denied.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - roles
      summary: Add a new role.
      description: Adds a new role to the collection.
      operationId: addRole
      requestBody:
        description: Group object that needs to be added to the collection
        required: true
        content:
          application/json:
            schema:
              type: object
          application/xml:
            schema:
              type: object
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
            application/xml:
              schema:
                type: object
        '400':
          description: Token missing or access denied.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/{roleId}:
    delete:
      tags:
        - roles
      summary: Deletes a role.
      description: Deletes a role by ID.
      operationId: deleteRole
      responses:
        '204':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
            application/xml:
              schema:
                type: object
        '400':
          description: Token missing or access denied.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/{roleId}/user:
    put:
      tags:
        - roles
      summary: Assigns a user to a role.
      description: Assigns a user to a role by ID.
      operationId: assignUserToRole
      responses:
        '204':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
            application/xml:
              schema:
                type: object
        '400':
          description: Token missing or access denied.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - roles
      summary: Removes a user from a role.
      description: Assigns a user to a role by ID.
      operationId: assignUserToRole
      responses:
        '204':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
            application/xml:
              schema:
                type: object
        '400':
          description: Token missing or access denied.
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...